modified:   lib/widgets/player_widget.dart 	renamed:    lib/widgets/music_player_progress_indicator.dart -> lib/widgets/progress_indicator.dart
